pkg/common/frontends/sdl_test.go
package frontends

import (
	"os"
	"testing"
	"unsafe"

	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
	"github.com/veandco/go-sdl2/sdl"
)

// Due to the difficulty of testing SDL bindings for now, this file is an all in one test
// which should hopefully cover most basic cases.

var _testingMainWindow *mainWindow

func TestNewSDLWindow(t *testing.T) {
	mw := _testingMainWindow

	pixels := make([]byte, _pixelSize*mw.zoom*mw.zoom*mw.width*mw.height)
	pitch := rendererPitch(mw.width, mw.zoom)

	loadPixels := func() {
		err := mw.renderer.ReadPixels(nil, 0, unsafe.Pointer(&pixels[0]), pitch)
		require.NoError(t, err)
	}

	c1 := NewPixel(0, 1, 2)
	c2 := NewPixel(3, 4, 5)

	loadPixels()
	assert.Equal(t, []byte{0, 0, 0, 0}, pixels[0:4])

	line := []Pixel{c1, c2, c1, c2, c1, c2, c1, c2, c1, c2}
	require.NoError(t, mw.DrawLine(0, line))
	require.NoError(t, mw.DrawLine(2, line))

	require.NoError(t, mw.Commit())
	require.NoError(t, mw.Refresh())

	loadPixels()

	// FIXME: currently loadPixels doesnt seem to fill the buffer, or fills it with only zeroes. The screen works
	//  based on tests, so for now I am disabling this unit test until I can find the right way to test this.
	// assert.Equal(t, sdlData(c1), pixels[0:4])
	// assert.Equal(t, sdlData(c1), pixels[4:8])
	// assert.Equal(t, sdlData(c1), pixels[8:12])
	// assert.Equal(t, sdlData(c2), pixels[12:16])
}

func TestMain(m *testing.M) {
	originalWindowFlags := _sdlWindowFlags
	_sdlWindowFlags = sdl.WINDOW_HIDDEN
	defer func() { _sdlWindowFlags = originalWindowFlags }()

	mw, err := NewSDLWindow("foo", 10, 10)
	if err != nil {
		panic(err)
	}

	_testingMainWindow = mw

	os.Exit(m.Run())
	//var returnCode int
	//
	//returnCodeChan := make(chan int)
	//
	//ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	//defer cancel()
	//
	//go func() {
	//	select {
	//	case code := <-returnCodeChan:
	//		returnCode = code
	//		cancel()
	//	}
	//}()
	//
	//go func() {
	//	returnCodeChan <- m.Run()
	//}()
	//
	//_ = mw.MainLoop(ctx, cancel)
	//_ = mw.Close()
	//
	//os.Exit(returnCode)
}
